Initial Commit
After much stumbling-in-the-darkness, this commit has a devshell that works in
providing a python environment with the non-nixpkgs dependencies in a sub-flake.
I'm not even sure if this is a good solution, but for now it stays.
It is annoying however to constantly pushd dependencies/ and nix flake update,
whenever making changes here....

I think the model of placing dependencies in an overlay, then merging overlays
together is a very pleasant way of composing packages. It seems to be a endorsed
style in the Nix community.

I am hoping this explicit-packaging of python modules will compose well with
mach-nix for specifying the python dependencies of ibex.